delete old scorecard pages
