+++ dotcom-rendering/src/components/CricketMiniMatchStats.tsx
@@ -8,9 +8,9 @@ import {
 import type { SWRConfiguration } from 'swr';
 import useSWR from 'swr';
 import { safeParse } from 'valibot';
-import type { CricketMatchStatsSummary } from '../cricketMatchV2';
-import { parseMatchStatsSummary } from '../cricketMatchV2';
-import { feCricketMatchStatsSummarySchema } from '../frontend/feCricketMatchPage';
+import type { CricketMatchStatsSummary } from '../cricketMatch';
+import { parseMatchStatsSummary } from '../cricketMatch';
+import { feCricketMatchStatsSummarySchema } from '../frontend/feCricketMatchData';
 import type { Result } from '../lib/result';
 import { error, fromValibot, ok } from '../lib/result';
 import { palette } from '../palette';
