--- dotcom-rendering/src/components/CricketScorecardNew.stories.tsx
@@ -1,377 +1,0 @@
-import { centreColumnDecorator } from '../../.storybook/decorators/gridDecorators';
-import { allModes } from '../../.storybook/modes';
-import preview from '../../.storybook/preview';
-import { CricketScorecardNew as CricketScorecardNewComponent } from './CricketScorecardNew';
-
-const meta = preview.meta({
-	component: CricketScorecardNewComponent,
-	title: 'Components/CricketScorecardNew',
-	decorators: [centreColumnDecorator],
-	parameters: {
-		chromatic: {
-			modes: {
-				'light leftCol': allModes['light leftCol'],
-			},
-		},
-	},
-});
-
-export const CricketScorecardNew = meta.story({
-	name: 'Cricket Scorecard (New)',
-	args: {
-		allInnings: [
-			{
-				description: 'India first innings',
-				battingTeam: 'India',
-				order: 1,
-				declared: false,
-				forfeited: false,
-				inningsTotals: {
-					runs: 254,
-					overs: '49.0',
-					extras: 8,
-					wickets: 7,
-				},
-				extras: {
-					byes: 1,
-					legByes: 2,
-					wides: 5,
-					noBalls: 0,
-					penalties: 0,
-				},
-				batters: [
-					{
-						name: 'Rohit Sharma',
-						ballsFaced: 83,
-						runs: 76,
-						fours: 7,
-						sixes: 3,
-						howOut: 'st Latham b Ravindra',
-						out: true,
-						onStrike: false,
-						nonStrike: true,
-					},
-					{
-						name: 'Shubman Gill',
-						ballsFaced: 50,
-						runs: 31,
-						fours: 0,
-						sixes: 1,
-						howOut: 'c Phillips b Santner',
-						out: true,
-						onStrike: false,
-						nonStrike: true,
-					},
-					{
-						name: 'Virat Kohli',
-						ballsFaced: 2,
-						runs: 1,
-						fours: 0,
-						sixes: 0,
-						howOut: 'lbw b Bracewell',
-						out: true,
-						onStrike: false,
-						nonStrike: true,
-					},
-					{
-						name: 'Shreyas Iyer',
-						ballsFaced: 62,
-						runs: 48,
-						fours: 2,
-						sixes: 2,
-						howOut: 'c Ravindra b Santner',
-						out: true,
-						onStrike: false,
-						nonStrike: true,
-					},
-					{
-						name: 'Lokesh Rahul',
-						ballsFaced: 45,
-						runs: 39,
-						fours: 3,
-						sixes: 0,
-						howOut: 'run out (Williamson)',
-						out: true,
-						onStrike: false,
-						nonStrike: true,
-					},
-					{
-						name: 'Hardik Pandya',
-						ballsFaced: 18,
-						runs: 22,
-						fours: 2,
-						sixes: 1,
-						howOut: 'c Bracewell b Smith',
-						out: true,
-						nonStrike: true,
-						onStrike: false,
-					},
-					{
-						name: 'Ravindra Jadeja',
-						ballsFaced: 24,
-						runs: 30,
-						fours: 1,
-						sixes: 0,
-						howOut: 'not out',
-						out: false,
-						onStrike: true,
-						nonStrike: true,
-					},
-				],
-				bowlers: [
-					{
-						name: 'Mohammed Shami',
-						overs: 9,
-						maidens: 0,
-						runs: 74,
-						wickets: 1,
-						balls: 55,
-					},
-					{
-						name: 'Hardik Pandya',
-						overs: 3,
-						maidens: 0,
-						runs: 30,
-						wickets: 0,
-						balls: 18,
-					},
-					{
-						name: 'Varun Chakaravarthy',
-						overs: 10,
-						maidens: 0,
-						runs: 45,
-						wickets: 2,
-						balls: 60,
-					},
-					{
-						name: 'Kuldeep Yadav',
-						overs: 10,
-						maidens: 0,
-						runs: 40,
-						wickets: 2,
-						balls: 60,
-					},
-					{
-						name: 'Axar Patel',
-						overs: 8,
-						maidens: 0,
-						runs: 29,
-						wickets: 0,
-						balls: 48,
-					},
-					{
-						name: 'Ravindra Jadeja',
-						overs: 10,
-						maidens: 0,
-						runs: 30,
-						wickets: 1,
-						balls: 60,
-					},
-				],
-				fallOfWickets: [
-					{
-						order: 1,
-						name: 'Shubman Gill',
-						runs: 105,
-					},
-					{
-						order: 2,
-						name: 'Virat Kohli',
-						runs: 106,
-					},
-					{
-						order: 3,
-						name: 'Rohit Sharma',
-						runs: 122,
-					},
-				],
-			},
-			{
-				description: 'New Zealand first innings',
-				battingTeam: 'New Zealand',
-				order: 2,
-				declared: false,
-				forfeited: false,
-				inningsTotals: {
-					runs: 254,
-					overs: '49.0',
-					extras: 8,
-					wickets: 7,
-				},
-				extras: {
-					byes: 0,
-					legByes: 3,
-					noBalls: 0,
-					penalties: 0,
-					wides: 13,
-				},
-				batters: [
-					{
-						name: 'Will Young',
-						ballsFaced: 23,
-						runs: 15,
-						fours: 2,
-						sixes: 0,
-						howOut: 'lbw b Vinod',
-						out: true,
-						onStrike: false,
-						nonStrike: true,
-					},
-					{
-						name: 'Rachin Ravindra',
-						ballsFaced: 29,
-						runs: 37,
-						fours: 4,
-						sixes: 1,
-						howOut: 'b Yadav',
-						out: true,
-						onStrike: false,
-						nonStrike: true,
-					},
-					{
-						name: 'Kane Williamson',
-						ballsFaced: 14,
-						runs: 11,
-						fours: 1,
-						sixes: 0,
-						howOut: 'c & b Yadav',
-						out: true,
-						onStrike: false,
-						nonStrike: true,
-					},
-					{
-						name: 'Daryl Mitchell',
-						ballsFaced: 101,
-						runs: 63,
-						fours: 3,
-						sixes: 0,
-						howOut: 'c Sharma b Ahmed',
-						out: true,
-						onStrike: false,
-						nonStrike: true,
-					},
-					{
-						name: 'Tom Latham',
-						ballsFaced: 56,
-						runs: 44,
-						fours: 4,
-						sixes: 0,
-						howOut: 'c Kohli b Shami',
-						out: true,
-						onStrike: false,
-						nonStrike: true,
-					},
-					{
-						name: 'Glenn Phillips',
-						ballsFaced: 34,
-						runs: 55,
-						fours: 3,
-						sixes: 3,
-						howOut: 'not out',
-						out: false,
-						onStrike: true,
-						nonStrike: true,
-					},
-					{
-						name: 'Michael Bracewell',
-						ballsFaced: 8,
-						runs: 5,
-						fours: 0,
-						sixes: 0,
-						howOut: 'not out',
-						out: false,
-						onStrike: false,
-						nonStrike: true,
-					},
-				],
-				bowlers: [
-					{
-						name: 'Mohammed Shami',
-						overs: 9,
-						maidens: 0,
-						runs: 74,
-						wickets: 1,
-						balls: 54,
-					},
-					{
-						name: 'Hardik Pandya',
-						overs: 3,
-						maidens: 0,
-						runs: 30,
-						wickets: 0,
-						balls: 18,
-					},
-					{
-						name: 'Varun Chakaravarthy',
-						overs: 10,
-						maidens: 0,
-						runs: 45,
-						wickets: 2,
-						balls: 60,
-					},
-					{
-						name: 'Kuldeep Yadav',
-						overs: 10,
-						maidens: 0,
-						runs: 40,
-						wickets: 2,
-						balls: 60,
-					},
-				],
-				fallOfWickets: [
-					{
-						order: 1,
-						name: 'Will Young',
-						runs: 57,
-					},
-					{
-						order: 2,
-						name: 'Rachin Ravindra',
-						runs: 69,
-					},
-				],
-			},
-		],
-		officials: [
-			'P R Reiffel',
-			'R K Illingworth',
-			'J S Wilson',
-			'H D P K Dharmasena',
-			'R S Madugalle',
-		],
-		homeTeam: {
-			paID: 'a1b2c3d4-e5f6-7890-abcd-ef1234567890',
-			name: 'India',
-			lineup: [
-				'Rohit Sharma',
-				'Shubman Gill',
-				'Virat Kohli',
-				'Shreyas Iyer',
-				'Axar Patel',
-				'Lokesh Rahul',
-				'Hardik Pandya',
-				'Ravindra Jadeja',
-				'Mohammed Shami',
-				'Kuldeep Yadav',
-				'Varun Chakaravarthy',
-			],
-		},
-		awayTeam: {
-			paID: 'b2c3d4e5-f6a7-8901-bcde-f12345678901',
-			name: 'New Zealand',
-			lineup: [
-				'Rachin Ravindra',
-				'Will Young',
-				'Kane Williamson',
-				'Daryl Mitchell',
-				'Tom Latham',
-				'Glenn Phillips',
-				'Michael Bracewell',
-				'Mitchell Santner',
-				'Nathan Smith',
-				'Kyle Jamieson',
-				"Will O'Rourke",
-			],
-		},
-	},
-});
