+++ dotcom-rendering/src/model/validate.ts
@@ -3,14 +3,12 @@ import type { Options } from 'ajv';
 import Ajv from 'ajv';
 import addFormats from 'ajv-formats';
 import type { FEArticle } from '../frontend/feArticle';
-import type { FECricketMatchPage } from '../frontend/feCricketMatchPage';
 import type { FEFootballMatchInfoPage } from '../frontend/feFootballMatchInfoPage';
 import type { FEFootballMatchListPage } from '../frontend/feFootballMatchListPage';
 import type { FEFootballTablesPage } from '../frontend/feFootballTablesPage';
 import type { FEFront } from '../frontend/feFront';
 import type { FETagPage } from '../frontend/feTagPage';
 import articleSchema from '../frontend/schemas/feArticle.json';
-import cricketMatchPageSchema from '../frontend/schemas/feCricketMatchPage.json';
 import footballMatchInfoPageSchema from '../frontend/schemas/feFootballMatchInfoPage.json';
 import footballMatchListPageSchema from '../frontend/schemas/feFootballMatchListPage.json';
 import footballTablesPageSchema from '../frontend/schemas/feFootballTablesPage.json';
@@ -50,9 +48,7 @@ const validateFootballMatchListPage = ajv.compile<FEFootballMatchListPage>(
 const validateFootballTablesPage = ajv.compile<FEFootballTablesPage>(
 	footballTablesPageSchema,
 );
-const validateCricketMatchPage = ajv.compile<FECricketMatchPage>(
-	cricketMatchPageSchema,
-);
+
 const validateFootballMatchInfoPage = ajv.compile<FEFootballMatchInfoPage>(
 	footballMatchInfoPageSchema,
 );
@@ -152,22 +148,6 @@ export const validateAsFootballTablesPage = (
 	throw new TypeError(
 		`Unable to validate request body for url ${url}.\n
             ${JSON.stringify(validateFootballMatchListPage.errors, null, 2)}`,
-	);
-};
-
-export const validateAsCricketMatchPageType = (
-	data: unknown,
-): FECricketMatchPage => {
-	if (validateCricketMatchPage(data)) return data;
-
-	const url =
-		isObject(data) && isObject(data.config) && isString(data.config.pageId)
-			? data.config.pageId
-			: 'unknown url';
-
-	throw new TypeError(
-		`Unable to validate request body for url ${url}.\n
-            ${JSON.stringify(validateCricketMatchPage.errors, null, 2)}`,
 	);
 };
